+++ src/libdinoseq/controllerinfo.hpp
@@ -31,16 +31,16 @@ namespace Dino {
   class ControllerInfo {
   public:
     
-    ControllerInfo(long number, const std::string& name);
+    ControllerInfo(uint32_t number, const std::string& name);
     
-    long get_number() const;
+    uint32_t get_number() const;
     int get_default() const;
     int get_min() const;
     int get_max() const;
     const std::string& get_name() const;
     bool get_global() const;
     
-    void set_number(long n);
+    void set_number(uint32_t n);
     void set_default(int def);
     void set_min(int min);
     void set_max(int max);
@@ -49,7 +49,7 @@ namespace Dino {
     
   protected:
     
-    long m_number;
+    uint32_t m_number;
     int m_default;
     int m_min;
     int m_max;
